% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204185.R
\name{add}
\alias{add}
\title{self coded regression}
\usage{
add(a, b)
}
\arguments{
\item{a}{0 to preform standard regression,1 for ridge regression and 2 for locally weighted regression}

\item{b}{lambda the multiple of identity matrix added to xTx in the process of ridge regression}
}
\value{
the coefficient of the regression whose last column is the value of intercept
}
\description{
apply standard,ridge or locally weighted regression(using Gaussian kernel) on a given data frame(the predicted variable has to be at the last column 
of the data frame like 'Boston' in MASS we used in vignettes )
}
\examples{
\dontrun{
coef <- add(1,2)
}
}
