% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transformPermCoding.R
\name{adjust_pattern}
\alias{adjust_pattern}
\title{A function to create new pattern-coding schemes for the Fisher Information.}
\usage{
adjust_pattern(target_pattern, ndemb)
}
\arguments{
\item{target_pattern}{A numeric matrix that specifies the pattern to be transformed into the position vector. ATTENTION: Pattern should be in the ranks permutation notation, otherwise does not really make sense.}

\item{ndemb}{Embedding dimension of the ordinal patterns (i.e. sliding window size). Should be chosen such as length(x) >> ndemb}
}
\value{
A numeric matrix that contains the permutation matrix.
}
\description{
Adjusts and reorders a pattern ordering matrix.
}
\details{
This function reorders permutations based on "jumps" or based on "bitflips".
}
\author{
Sebastian Sippel
}
\references{
Sebastian Sippel (2014). Master Thesis. University of Bayreuth.
}

