% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{Algorithm1}
\alias{Algorithm1}
\title{Algorithm to get a coefficient beta in compositional covariates problem}
\usage{
Algorithm1(x, y, learning_rate, convergence_limit = 1e-05)
}
\arguments{
\item{x}{n x p design matrix with sample size n and dimensionality p}

\item{y}{n-vector of response}

\item{learning_rate}{regularization parameter of L1 penalty controling its maximum concavity}

\item{convergence_limit}{tolerance for updating.  Default = 1e-5}
}
\value{
a coeffcient value in compositional covariates problem
}
\description{
Core pragram of the algorithm (Lin, W., Shi, P., Feng, R., Li, H, 2014), 
which solves the problem of variable selection in regression with compositional covariates.
}
