% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCtestforAR.R
\name{ar}
\alias{ar}
\title{generate AR(1)}
\usage{
ar(phi, t, y0)
}
\arguments{
\item{phi}{unite root}

\item{t}{time}

\item{y0}{orignal data}
}
\value{
a time series
}
\description{
generate AR(1)
}
\examples{
\dontrun{
data=ar(0.6,10,rep(0,5))
print(data)
}
}
