% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmark.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{system.time} to compare the performance of C functions (\code{fibC} and \code{fibR}).
}
\examples{
\dontrun{
system.time(fibR(35))
system.time(fibC(35))
}
}
