% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{boot_ci}
\alias{boot_ci}
\title{Compute bootstrap confidence intervals for the mean by the standard normal, basic, percentile, and BCa methods.}
\usage{
boot_ci(x)
}
\arguments{
\item{x}{a numeric vector of data sample}
}
\value{
a list contains the 4 C.I. and other information.
}
\description{
Compute 95 bootstrap confidence intervals for the mean by the standard normal, basic, percentile, and BCa methods.
}
\examples{
\dontrun{
data(aircondit)
aircondit <- as.matrix(aircondit)
bo <- boot_ci(aircondit)
print(bo)
}
}
