% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottesttwo.R
\name{boottesttwo}
\alias{boottesttwo}
\title{Two-sample bootstrap test}
\usage{
boottesttwo(x, y, b)
}
\arguments{
\item{x}{vector containing first sample.}

\item{y}{vector containing second sample.}

\item{b}{number of bootstrap replications}
}
\value{
origtest: value of test statistic on original samples

pvalue: bootstrap p-value

teststatall: vector of bootstrap test statistics
}
\description{
This function is used for the two-sample bootstrap test,
where the test statistic is the difference between the two sample means
}
\examples{
x=rnorm(15,2,4)
y=rnorm(10,4,4)
boottesttwo(x,y,50)
}
