% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CI.R
\name{calculate_CI}
\alias{calculate_CI}
\title{Confidence interval}
\usage{
calculate_CI(x, conf = 0.95)
}
\arguments{
\item{x}{vector, data of length N}

\item{conf}{numeric, alpha}
}
\value{
confidence interval
}
\description{
Confidence interval
}
\details{
Takes as input a vector of data of length N and alpha and calculates a confidence interval for the estimate of the mean in the populaiton.
}
\examples{
x <- rnorm(1:10)
calculate_CI(x, 0.9)

}
