% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainCauchy.R
\name{cauchy.chain}
\alias{cauchy.chain}
\title{The MC chain of Cauchy distribution}
\usage{
cauchy.chain(n, sigma, x0, location, scale)
}
\arguments{
\item{n}{the length of the chain.}

\item{sigma}{The standard deviation of the proposed distribution.}

\item{x0}{The initial value of the chain.}

\item{location, scale}{location and scale parameters.}
}
\value{
\code{x}  return the chain.
\code{k}  Returns the number of rejections.
}
\description{
Give a MC chain of Cauchy distribution
with location parameter \code{location} and scale parameter \code{scale}.
}
\details{
If \code{location} or \code{scale} are not specified, they assume the
default values of 0 and 1 respectively.

The Cauchy distribution with location \code{l} and scale \code{s} has density
\code{f(x)=1/(pi s(1+((x-l)/s)^2))} for all \code{x}.
}
\examples{
\dontrun{
n<-10000
cauchy<-cauchy.chain(n,sigma=1,x0=0,location=0,scale=1)
cauchy$k/n
hist(cauchy$x[1001:n],freq=FALSE,main="cauchy瀵嗗害鍥戼㹥",breaks=60)
curve(dcauchy(x,0,1),add=TRUE)
}
}
