% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{cfint}
\alias{cfint}
\title{CIdichotomy}
\usage{
cfint(lam, alpha)
}
\arguments{
\item{lam}{the confidence coefficient}

\item{alpha}{the shape parameter of the gamma function}
}
\value{
a label for test set
}
\description{
Find confidence intervals by dichotomy
}
