% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/changePermCodingOPD.R
\name{changePermCodingOPD}
\alias{changePermCodingOPD}
\title{A function to change the ordering scheme of ordinal patterns}
\usage{
changePermCodingOPD(opd, ndemb, target_pattern, use_target_pattern = T, transform_vec = NA)
}
\arguments{
\item{opd}{A numeric vector that comprises an ordinal pattern distribution in the Keller coding scheme (Physica A 356 (2005) 114-120).}

\item{ndemb}{Embedding dimension of the ordinal patterns (i.e. sliding window size).}

\item{target_pattern}{If specified, must consist of a numeric matrix of the dimensions c(factorial(ndemb), ndemb), which consists of a user-specified permutation coding scheme ('ranks-based').}

\item{transform_vec}{If specified, a numeric vector that contains indices to change an ordinal pattern distribution from the (original) coding scheme to any user-specified permutation coding scheme.}
}
\value{
An ordinal pattern distribution vector in a user-specified permutation coding scheme.
}
\description{
Changes the permutation ordering scheme of a ordinal pattern distribution. This affects the computation of local information measures, but is insensitive to global measures.
}
\details{
This function returns an ordinal pattern distribution in a user-specified permutation coding scheme.
}
\examples{
x = arima.sim(model=list(ar = 0.3), n = 10^4)
opd = ordinal_pattern_distribution(x = x, ndemb = 6)
}
\author{
Sebastian Sippel
}
\references{
none
}

