% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count}
\alias{count}
\title{count loss function}
\usage{
count(x, y, beta, i)
}
\arguments{
\item{x}{Independent variable (dataframe)}

\item{y}{Dependent variable (dataframe)}

\item{beta}{coefficent}

\item{i}{ith unit}
}
\value{
The loss function for the i-th unit
}
\description{
count loss function
}
\examples{
\dontrun{
x <- matrix(1:9, nrow = 3, ncol = 3)
y <- matrix(10:18, nrow = 3, ncol = 3)
count(x, y, 1, 1)
}
}
