% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{count5test}
\alias{count5test}
\title{Equal variance test for two samples(equal size)}
\usage{
count5test(x, y)
}
\arguments{
\item{x}{the one sample vector}

\item{y}{the other sample vector}
}
\value{
the value 1(reject H0) or 0 (do not reject H0)
}
\description{
It is applied for independent random samples when the random variables are similarly 
distributed and sample sizes are equal. For two independent random samples of 
equal size, the maximum number of extreme observation values is calculated and compared with 5.  
The null hypothesis is equal variance.
}
\examples{
\dontrun{
x<-rnorm(20,0,2)
y<-rnorm(20,1,4)
count5test(x,y)
}
}
