% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{count5test_permutation}
\alias{count5test_permutation}
\title{Count five permutation test}
\usage{
count5test_permutation(z, R)
}
\arguments{
\item{z}{a numeric vector which is generated by combining sample x and sample y}

\item{R}{the permutation times}
}
\value{
a number in [0,1] which represents the frequency of count5test being rejected.
}
\description{
Implement a permutation test for equal variance
based on the maximum number of extreme points that applies
when sample sizes are not necessarily equal.
}
\examples{
\dontrun{
n1 <- 20; n2 <-50
mu1 <- mu2 <- 0
sigma1 <- sigma2 <- 1
m <- 1e3
alphahat <- mean(replicate(m,expr={
x=rnorm(n1,mu1,sigma1)
y=rnorm(n2,mu2,sigma2)
z=c(x,y)
count5test_permutation(z,1000)}))
print(alphahat)
}
}
