% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CURE.R
\name{CURE}
\alias{CURE}
\title{CURE using Lars algorithm}
\usage{
CURE(yk, X, d0, u0, v0, lambda, mu, eps, max_iter)
}
\arguments{
\item{yk}{the vector of current response variable}

\item{X}{design matrix}

\item{d0}{initial value of d}

\item{u0}{initial value of u}

\item{v0}{initial value of v}

\item{lambda}{tuning parameter}

\item{mu}{parameter}

\item{eps}{iteration residual bound}

\item{max_iter}{maximal number of iteration}
}
\value{
a list including estimated coefficients (d,u,v) and iteration
}
\description{
CURE using Lars algorithm
}
\details{
This is some descriptio of this function.
}
