% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVM.test.R
\name{CVM.test}
\alias{CVM.test}
\title{Two-sample Cramer-von Mises test.}
\usage{
CVM.test(x, y)
}
\arguments{
\item{x}{the first sample.}

\item{y}{the second sample.}
}
\value{
a named list with W0 and p-value of permutation test.
}
\description{
Test whether two samples have same distributuion. \code{CVM.test} will calculate the value of statistic W and return as a list (W0, p-value).
}
\examples{

CVM.test(rnorm(10), rnorm(20))

}
