% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DAP}
\alias{DAP}
\title{A multi-prob calculator for Deliveryman Assignment Problem}
\usage{
DAP(x, target, n)
}
\arguments{
\item{x}{the probability vector of each deliveryman}

\item{target}{there are excatly target number of deliverymen take order}

\item{n}{the length of vector x}
}
\value{
the objective probability
}
\description{
A multi-prob calculator using Rcpp
}
\examples{
\dontrun{
prob <- c(0.5, 0.5, 0.5, 0.5, 0.5)
target <- 1
n <- 5
DAP(prob, target, n)
}
}
