\name{data}
\alias{data}

\title{
A simulated dataset used for illustration of functions 'vaccR' and 'vaccC'.
}
\description{
The dataset includes three variables (age, female, ily) of length 1000, which is used to illustrate the function 'vaccR' and 'vaccC'.
}
\usage{
data(data)
}
\format{
  A data frame with 1,000 observations on the following 3 variables.
  \describe{
    \item{\code{age}}{age (numeric)}
    \item{\code{female}}{gender (TRUE = female, FALSE = male)}
    \item{\code{ily}}{ily (TRUE or FALSE)}
  }
}

\keyword{datasets}
