% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{data_generate}
\alias{data_generate}
\title{Generate the data for numeric study}
\usage{
data_generate(n, p, beta, rho = 0.2, sig = 0.5)
}
\arguments{
\item{n}{sample size n}

\item{p}{dimension}

\item{beta}{true coefficient beta}

\item{rho}{for generate covariance matrix. Default =0.2}

\item{sig}{covariance for white noise. Default = 0.5}
}
\value{
a list of data generated including data X and response y in the form of matrix
}
\description{
generate the data for numeric study in the exact paper and for example.
}
\examples{
\dontrun{
p=30
beta = c(c(1,-0.8,0.6,0,0,-1.5,-0.5,1.2),rep(0,p-8))
data1 = data_generate(n=50,p=30,beta = beta)
result = lrate_GIC(X=data1[[1]],y=data1[[2]])
learning_rate_GIC = result[1]
GIC_value = result[2]

#estimate beta
beta_est = estimate(X=data1[[1]],y=data1[[2]],learning_rate = unlist(learning_rate_GIC))

#compute estimation error.
pred_error(x=data1[[1]],y=data1[[2]],beta_hat = beta_est)

#compute estimation accuracy. 
est_acc(beta_est,beta)
}
}
