% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA24204168R.R
\name{dif}
\alias{dif}
\title{Perform time differencing on panel data}
\usage{
dif(x, y)
}
\arguments{
\item{x}{Independent variable (dataframe)}

\item{y}{Dependent variable (dataframe)}
}
\value{
Differenced independent variable (x) and dependent variable (y)
}
\description{
Perform differencing on each panel data unit by time cross-section to eliminate the effect of the intercept term
}
\examples{
\dontrun{
x <- data.frame(matrix(1:9, nrow = 3, ncol = 3))
y <- data.frame(matrix(10:18, nrow = 3, ncol = 3))
dif(x, y)
}
}
