% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204185.R
\name{distribute}
\alias{distribute}
\title{distribute algorithm}
\usage{
distribute(matrices_get)
}
\arguments{
\item{matrices_get}{the list of input matrices}
}
\value{
the common right singular vectors of input matrices
}
\description{
Aiming to recover the common right singular vectors of the input matrix columns.
}
