% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204185.R
\name{distribute_monte}
\alias{distribute_monte}
\title{simulation}
\usage{
distribute_monte(m, n, p, monte)
}
\arguments{
\item{m}{Number of sub-servers}

\item{n}{data volume on each server}

\item{p}{missing probability p}

\item{monte}{number of Monte Carlo iterations}
}
\value{
the common right singular vectors of input matrices
}
\description{
Directly encapsulating the numerical simulation process, you can obtain the distance rho 
between the differences in the right singular vectors before and after the response by simply inputting the required parameters.
}
\examples{
\dontrun{
distribute_monte(m = 10, n = 100, p = 0.1, monte = 10)
}
}
