% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rayleigh.R
\name{dRayleigh}
\alias{dRayleigh}
\alias{pRayleigh}
\alias{qRayleigh}
\alias{rRayleigh}
\title{The Reyleigh distribution}
\usage{
dRayleigh(x,sigma)
pRayleigh(q,sigma,lower.tail=TRUE)
qRayleigh(p,sigma,lower.tail=TRUE)
rRayleigh(n,sigma)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{sigma}{the scale parameter of Reyleigh distribution.}

\item{p}{vector of probabilities.}

\item{n}{the number of samples.}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P[X<=x]}, otherwise, \code{P[X > x]}.}
}
\value{
dt gives the density, pt gives the distribution function, qt gives
 the quantile function, and rt generates random deviates.
}
\description{
Density, distribution function, quantile function and
 random generation for the Reyleigh distribution.
}
\examples{
\dontrun{
n<-10000
y<-rRayleigh(n,sigma=2)
hist(y,breaks=seq(0,ceiling(max(y)),0.1),freq = FALSE,
main = "Histogram of Reyleigh",xlab = " ")
x<-seq(0,8.5,0.05)
lines(x,dRayleigh(x,sigma=2),col="red")
qRayleigh(0.05,sigma=2)
pRayleigh(0.6405828,sigma = 2)
}
}
