% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIVE-R.R
\name{EigenvalueRatio}
\alias{EigenvalueRatio}
\title{A Estimator of the number of factors using Eigenvalue Ratio method}
\usage{
EigenvalueRatio(X)
}
\arguments{
\item{X}{the data matrix}
}
\value{
Estimate of the number of factors
}
\description{
A Estimator of the number of factors using Eigenvalue Ratio method
}
\examples{
\dontrun{
X <- mvrnorm(n, rep(0, p), Sigma)
khat <- EigenvalueRatio(X)
}
}
