% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{EM_blood}
\alias{EM_blood}
\title{Use EM algorithm to solve a MLE problem about A-B-O blood type}
\usage{
EM_blood(n.A, n.B, nOO, nAB)
}
\arguments{
\item{n.A}{the number of samples which are A-type.}

\item{n.B}{the number of samples which are B-type.}

\item{nOO}{the number of samples which are O-type.}

\item{nAB}{the number of samples which are AB-type.}
}
\value{
a list contains a matrix r composed by the sequences of p and q value, and a numeric vector MLE which is their correponding log-maximum likelihood.
}
\description{
Use EM algorithm to solve MLE of p and q in A-B-O blood type problem condsidering missing data nAA and nBB.
}
\examples{
\dontrun{
n.A <- 444; n.B <- 132; nOO <- 361; nAB <- 63
t <- EM_blood(n.A,n.B,nOO,nAB)
t$r
plot(-t$mle,type='l')
}
}
