% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{epan}
\alias{epan}
\title{Epanechnikov kernel function}
\usage{
epan(x)
}
\arguments{
\item{x}{real number}
}
\value{
Epanechnikov density(value) at x
}
\description{
As shown in the expression. It is also a density function.
}
\examples{
\dontrun{
epan(0)
}
}
