% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{eqdist.nn}
\alias{eqdist.nn}
\title{Equal Distribution Test by Nearest Neighbor Method}
\usage{
eqdist.nn(z, sizes, k, R)
}
\arguments{
\item{z}{A data set composed of two multidimensional samples}

\item{sizes}{A vector of two sample sizes}

\item{k}{the number of chosen neighbors}

\item{R}{number of bootstrap}
}
\value{
original nearest neighbor statistic and equal distribution test p-value
}
\description{
Use Bootstrap permutation method to obtain the sample of statistics
 and the p value for equal distribution test
}
\examples{
\dontrun{
n1<-10;n2<-20;n<-n1+n2
sizes<-c(n1,n2);k=3
m <- 500;
p.values<-numeric[m]
p.values <- matrix(NA,m,3)
for(i in 1:m){
x<-rnorm(10,0,1)
y<-rnorm(20,1,2)
z <- c(x, y)
p.values[i] <- eqdist.nn(z,sizes,k)$p.value
alpha <- 0.1;
pow <- mean(p.values<alpha)
}
}
}
