% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{eqtest.permu}
\alias{eqtest.permu}
\title{Permutation test for equal variance for two samples.}
\usage{
eqtest.permu(x, y)
}
\arguments{
\item{x}{the one sample vector}

\item{y}{the other sample vector}
}
\value{
permutation test p-value
}
\description{
A permutation test for equal variance based on the maximum number
 of extreme points that applies when sample sizes are not necessarily equal. 
 Calculate original test statistic and permutation test statistics, and obtain the 
 permutation test p-value defined as proportion of permutation test statistics 
 larger than the original statistic.
}
\examples{
\dontrun{
set.seed(11)
b<-numeric(1000)
for(i in 1:1000){
x<--rnorm(15,0,1)
y<-rnorm(20,1,2)
b[i]<-eqtest.permu(x,y)<=0.05
}
print(mean(b))
}
}
