% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{est_acc}
\alias{est_acc}
\title{Evaluate the performance of estimation}
\usage{
est_acc(beta_hat, beta_true)
}
\arguments{
\item{beta_hat}{estimated coefficient beta}

\item{beta_true}{true coefficient beta}
}
\value{
a list of loss_1,loss_2 and loss_infinity under different norms
}
\description{
compute the accuracy of estimation.
}
\examples{
\dontrun{
p=30
beta = c(c(1,-0.8,0.6,0,0,-1.5,-0.5,1.2),rep(0,p-8))
data1 = data_generate(n=50,p=30,beta = beta)
result = lrate_GIC(X=data1[[1]],y=data1[[2]])
learning_rate_GIC = result[1]
GIC_value = result[2]

#estimate beta
beta_est = estimate(X=data1[[1]],y=data1[[2]],learning_rate = unlist(learning_rate_GIC))

#compute estimation error.
pred_error(x=data1[[1]],y=data1[[2]],beta_hat = beta_est)

#compute estimation accuracy. 
est_acc(beta_est,beta)
}
}
