% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{esti_multi}
\alias{esti_multi}
\title{Kernel estimation of binary joint density}
\usage{
esti_multi(x, y, hx = 0.5, hy = 0.5, mat, method = c("multi", "sphe"))
}
\arguments{
\item{x}{the x-coordinate of point at which the joint density is to be obtained}

\item{y}{the y-coordinate of point at which the joint density is to be obtained}

\item{hx}{the bandwidth on the x-component}

\item{hy}{the bandwidth on the y-component}

\item{mat}{sample data}

\item{method}{the kernel we choose to estimate the joint pdf}
}
\value{
the joint density at point (x,y)
}
\description{
Multiplicative kernel and spherical-symmetric kernel can be selected 
to estimate the joint pdf of two variables according to their binary sample matrix.
}
\examples{
\dontrun{
attach(faithful)
esti_multi(x=2,y=56,mat=mat1,method = "multi")
xlab<-seq(0,6,0.1)
ylab<-seq(40,100,1)
z_density<-matrix(0,length(xlab),length(ylab))
for(i in 1:length(xlab)){
for(j in 1:length(ylab)){
 z_density[i,j]<-esti_multi(x=xlab[i],y=ylab[j],mat=mat1,method="multi")
 }}
 persp(xlab, ylab,z_density,phi = 30, theta = 40, col ="blue", border = 0, 
  zlim = c(0, 0.08),zlab = "Multiplicative KDE")
}
}
