% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuclideanDistanceSimilarity.R
\name{EuclideanDistanceSimilarity}
\alias{EuclideanDistanceSimilarity}
\title{The Euclidean Distance Similarity of Users}
\usage{
EuclideanDistanceSimilarity(M)
}
\arguments{
\item{M}{the data model}
}
\value{
s the similarity matrix of users
}
\description{
This function computes the euclidean distance similarity of users.
}
\examples{
\dontrun{
data(M)
S <- EuclideanDistanceSimilarity(M)
S
}
}
