% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwMetropolis.R
\name{eval_f0}
\alias{eval_f0}
\title{Mle for A-B-O blood type problem}
\usage{
eval_f0(x, x1, n.A = 444, n.B = 132, noo = 361, nAB = 63)
}
\arguments{
\item{x}{p is probability of A and q is probability of B}

\item{x1}{p1 is probability of A and q1 is probability of B}

\item{n.A}{number of n.A}

\item{n.B}{number of n.B}

\item{noo}{number of noo}

\item{nAB}{number of nAB}
}
\value{
mle
}
\description{
calculate the corresponding log-maximum likelihood values (for observed data)
}
