% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.R
\name{Exp}
\alias{Exp}
\title{Exponential transformation}
\usage{
Exp(x, k)
}
\arguments{
\item{x}{numeric, the number to be exponentiated}

\item{k}{numeric, the number of terms to be used in the series expansion beyond the constant 1}
}
\value{
the exponential of a number, numeric
}
\description{
Exponential transformation
}
\details{
This function computes the exponential of a number using the Taylor series expansion.
}
\examples{
Exp(2, 4)

}
