% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fibC}
\alias{fibC}
\title{Fibonacci sequence}
\usage{
fibC(n)
}
\arguments{
\item{n}{item \code{n}}
}
\value{
The value of item \code{n} of the Fibonacci sequence
}
\description{
Compute item \code{n} of the Fibonacci sequence using c++
}
\examples{
\dontrun{
fibC(29)
}
}
