% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fib.R
\name{fibR}
\alias{fibR}
\title{Fibonacci sequence}
\usage{
fibR(n)
}
\arguments{
\item{n}{item \code{n}}
}
\value{
The value of item \code{n} of the Fibonacci sequence
}
\description{
Compute item \code{n} of the Fibonacci sequence using R
}
\examples{
\dontrun{
fibR(29)
}
}
