% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complexity_measures.R
\name{fis}
\alias{fis}
\title{A (low-level) function to compute the Fisher-information}
\usage{
fis(odp)
}
\arguments{
\item{opd}{A numeric vector that details an ordinal pattern distribution in a user-specified permutation coding scheme.}
}
\value{
The normalized Fisher information measure in the range [0; 1].
}
\description{
The function computes the Fisher information, i.e. a local information measure based on two different discretizations.
}
\details{
The Fisher information is a local information and complexity measure, computed based on the ordinal pattern distribution.
The Fisher information is based on local gradients, hence it is sensitive to the permutation coding scheme.
Options for discretization: "Olivares.2012" or "Ferri.2009", following Fisher Information discretization schemes in the respective publications.
}
\examples{
x = arima.sim(model=list(ar = 0.3), n = 10^4)
opd = ordinal_pattern_distribution(x = x, ndemb = 6)
fis(opd = opd)
}
\author{
Sebastian Sippel
}
\references{
Olivares et al (2012): Physica A 391 (2012) 2518–2526; Olivares et al (2012): Physics Letters A 376 (2012) 1577-1583; Ferri et al (2009): Phys. Lett. A 373 (2009) 2210–2214.
}

