% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complexity_measures.R
\name{fisher_information}
\alias{fisher_information}
\title{A function to compute the Fisher-information}
\usage{
fisher_information(odp, x = NA, ndemb = NA, PatternCoding = "Default")
}
\arguments{
\item{opd}{A numeric vector that details an ordinal pattern distribution in a user-specified permutation coding scheme.}

\item{x}{(OPTIONAL) If opd is not specified, a time series vector x must be specified}

\item{ndemb}{(OPTIONAL) If x is given, the embedding dimension (ndemb) is required.}

\item{PatternCoding}{Character vector that specifies the pattern coding to use for calculating the Fisher Information. If "Default" is given, the Fisher Information is computed based on the (given) ordinal pattern distribution. Else, any pattern coding scheme can be given.
Logical. If TRUE, the Fisher information is computed based on the given ordinal pattern distribution. If FALSE, the ordinal pattern distribution must be given in the default coding scheme (i.e. Keller and Sinn, 2005), and the Fisher information based on various pattern coding schemes is returned.}
}
\value{
The normalized Fisher information measure in the range [0; 1]. If PatternCoding is specified, returns the Fisher information measure based on the specified permutation coding scheme.
}
\description{
The function computes the Fisher information, i.e. a local information measure.
}
\details{
The Fisher information is a local information and complexity measure, computed based on the ordinal pattern distribution.
The Fisher information is based on local gradients, hence it is sensitive to the permutation coding scheme.
}
\examples{
x = arima.sim(model=list(ar = 0.3), n = 10^4)
opd = ordinal_pattern_distribution(x = x, ndemb = 6)
fisher_information(opd = opd)
}
\author{
Sebastian Sippel
}
\references{
Olivares et al (2012): Physica A 391 (2012) 2518–2526
}

