% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs.R
\name{fitmodel}
\alias{fitmodel}
\title{analysis of the result based on SIRS Algorithm}
\usage{
fitmodel(A, y, res, beta)
}
\arguments{
\item{A}{n x p design matrix with sample size n and dimensionality p}

\item{y}{n-vector of response}

\item{res}{result of the \code{sirs}}

\item{beta}{the true coefficient. Default is rep(zero,p)(unknown).}
}
\value{
A list contain the number of null zero variables, the index of null zero variables, 
the estimate error, if the true beta is input, return recovery=1 while the estimated 
index of null zero variables is true, else return recovery=0.
}
\description{
A function to analysis the result of the function (\code{sirs}), including 
the index of null zero variables,the number of null zero variables, the estimate error, 
and wheather the estimator is successful recover the true coeffcient when the true beta is input.
when the true beta is input, set recovery=1 while the estimated 
index of null zero variables is true, else recovery=0.
}
\examples{
\dontrun{
library(MASS)
n = 100
p = 50
a = 0.4
rho = 0.5
mu = c(rep(0,p))
sigma1 <- matrix(0, p, p)
sigma1 <- rho ^ (abs(row(sigma1) - col(sigma1)))
A = mvrnorm(n,mu,sigma1)
y = A\%*\%beta
beta = c(0.5,-0.5,1,-1.2,-1,rep(0,p-5))
res = sirs(A,y,a)
obj = fitmodel(A,y,res,beta)
}
}
