% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Floyd.R
\name{floyd}
\alias{floyd}
\title{Using Floyd algorithm to find shortest path.}
\usage{
floyd(A)
}
\arguments{
\item{A}{The graph with weight.(numeric)}
}
\value{
the shortest cost and the path.
}
\description{
Floyd algorithm to find shortest path using R.
}
\examples{
\dontrun{
a<-matrix(0,7,7)
a[1,2]=3;a[1,3]=5; a[2,3]=1;a[2,4]=5;a[2,5]=8
a[3,4]=7;a[3,5]=4;a[3,6]=10; a[4,5]=3;a[4,7]=6
a[5,6]=1;a[5,7]=2; a[6,7]=2
b<-a+t(a)
floyd(b)
}
}
