% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{ftest}
\alias{ftest}
\title{F test of equal variance.}
\usage{
ftest(x, y, level = 0.055, n = 20)
}
\arguments{
\item{x}{the one sample vector}

\item{y}{the other sample vector}

\item{level}{Significant level}

\item{n}{Sample size}
}
\value{
the value 1(reject H0) or 0 (do not reject H0)
}
\description{
Under the null hypothesis (H0), the variance ratio follows the F-distribution.
Large values will reject the null hypothesis.
}
\examples{
\dontrun{
x<-rnorm(20,0,2)
y<-rnorm(20,1,4)
ftest(x,y)
}
}
