% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR20097.R
\name{gammaCI}
\alias{gammaCI}
\title{gammaCI}
\usage{
gammaCI(alpha, lambda, beta)
}
\arguments{
\item{alpha}{the confidence coefficient}

\item{lambda}{the shape parameter of the gamma function}

\item{beta}{the scale parameter of the gamma function}
}
\value{
confidence interval of gamma function
}
\description{
Find confidence intervals of gamma function by dichotomy
}
