% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR20097.R
\name{gammasolution}
\alias{gammasolution}
\title{gammasolution}
\usage{
gammasolution(value, lambda, beta, prec)
}
\arguments{
\item{value}{function value}

\item{lambda}{the shape parameter of t he gamma function}

\item{beta}{the scale parameter of the gamma function}

\item{prec}{The difference between the true value and the estimated value}
}
\description{
Find the solution of the function by dichotomy
}
