% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gaussian}
\alias{Gaussian}
\title{A Gaussian distribution random number generater}
\usage{
Gaussian(mu, sigma)
}
\arguments{
\item{mu}{the mean of Gaussian distribution}

\item{sigma}{the standard deviation of Gaussian distribution}
}
\value{
a random number form N(mu,sigma^2)
}
\description{
A Gaussian distribution random number generater using Rcpp
}
\examples{
\dontrun{
Guassian(0,1)
}
}
