% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwMetropolis.R
\name{Gelman.Rubin}
\alias{Gelman.Rubin}
\title{GelmanRubin}
\usage{
Gelman.Rubin(psi)
}
\arguments{
\item{psi}{psi is the statistic}
}
\value{
G-R statistic
}
\description{
Use the Gelman-Rubin method to monitor convergence of the chain
}
