% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transformPermCoding.R
\name{generate_lehmerperm_matrix}
\alias{generate_lehmerperm_matrix}
\title{A function to generate the Lehmer permutation ordering.}
\usage{
generate_lehmerperm_matrix(ndemb)
}
\arguments{
\item{ndemb}{The embedding dimension.}
}
\value{
A numeric matrix that contains the Lehmer permutation pattern.
}
\description{
Generates all permutations of a given embedding dimension, ordered according to the Lehmer coding scheme.
}
\details{
This function converts ranks to indices and back.
}
\author{
Sebastian Sippel
}
\references{
http://www.keithschwarz.com/interesting/code/?dir=factoradic-permutation
}

