% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/generateCodingScheme.R
\name{generateCodingScheme}
\alias{generateCodingScheme}
\title{A function to generate a variety of permutation coding schemes}
\usage{
generateCodingScheme(target_pattern = "lehmerperm", ndemb)
}
\arguments{
\item{target_pattern}{A character vector that specifies the pattern to be generated. Options are:
\itemize{
 \item{"lehmerperm"}
 \item{"kellerperm"}
 \item{"lehmerperm_Olivares"}
 \item{"kellerperm_Olivares"}
 \item{"lehmerperm_bitflips_adjusted"}
 \item{"kellerperm_bitflips_adjusted"}
 \item{"lehmerperm_jumps_adjusted"}
 \item{"kellerperm_jumps_adjusted"}
}}

\item{ndemb}{Embedding dimension of the ordinal patterns (i.e. sliding window size). Should be chosen such as length(x) >> ndemb}
}
\value{
A numeric matrix of the dimensions c(factorial(ndemb), ndemb), which contains all permutations.
}
\description{
Calculates a user-specified permutation coding scheme (see Olivares et al 2012 for details).
}
\details{
This function returns a numeric matrix of the permutation ordering scheme in the 'ranks' ordering scheme (i.e. permutations are denoted differently as in Keller and Sinn, 2005).
}
\examples{
generateCodingScheme(target_pattern = "lehmerperm", ndemb = 4)
}
\author{
Sebastian Sippel
}
\references{
see e.g. Olivares et al. 2012
}

