% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gibbs_binorm}
\alias{Gibbs_binorm}
\title{Gibbs_binorm}
\usage{
Gibbs_binorm(N, sigma1, sigma2, mu1, mu2, rho, s1, s2, X)
}
\arguments{
\item{N}{a matrix with dim n*p}

\item{sigma1}{part of variance matrix Sigma, position [1,1]}

\item{sigma2}{part of variance matrix Sigma, position [2,2]}

\item{mu1}{part of mean vector, position [1]}

\item{mu2}{part of mean vector, position [2]}

\item{rho}{part of variance matrix Sigma, position [1,2] or [2,1]}

\item{s1}{s1 = sqrt(1-rho^2)*sigma1}

\item{s2}{s2 = sqrt(1-rho^2)*sigma2}

\item{X}{initial matrix with first row given for the mean of giving start of Gibbs sampler}
}
\value{
a matrix of the result of Gibbs sampler of binorm situation
}
\description{
Implemetation of Gibbs sampler on binorm situation.
}
\examples{
\dontrun{
#initialize constants and parameters
N = 5000 #length of chain
burn = 1000 #burn-in length
X = matrix(0, N, 2) #the chain, a bivariate sample
rho = 0.9 #correlation
mu1 = 0
mu2 = 0
sigma1 = 1
sigma2 = 1
s1 = sqrt(1-rho^2)*sigma1
s2 = sqrt(1-rho^2)*sigma2
# generate the chain
X[1, ] = c(mu1, mu2) #initialize
set.seed(1)
X = Gibbs_binorm(N,sigma1,sigma2,mu1,mu2,rho,s1,s2,X)
b = burn + 1
x = X[b:N, ]
plot(x, main="", cex=.5, xlab=bquote(X[t]),ylab=bquote(Y[t]), ylim=range(x[,2]))
}
}
