% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{Gibbs_samp}
\alias{Gibbs_samp}
\title{Gibbs sampler for Bivariate distribution}
\usage{
Gibbs_samp(n = 3000, x1, x2, plot = c("T", "F"), paras = c(1, 2, 1, 2, 0.5))
}
\arguments{
\item{n}{The desired sample size}

\item{x1}{initial value of one variable}

\item{x2}{initial value of one variable}

\item{plot}{whether to generate scatter diagram of sample}

\item{paras}{a vector which includes means, standard deviations and the correlation coefficient}
}
\value{
A list that includes sample chain and mean, covariance matrix and correlation coefficient
}
\description{
In each iteration, samples at specific component locations are 
taken one by one according to the conditional distribution under other component samples.
In order to obtain the sanple from the goal distribution, discard the previous 
part of the unstable sample
}
\examples{
\dontrun{
out1<-Gibbs_samp(3000,0,0)
out1
out2<-Gibbs_samp(1000,0,0,plot="F")
out2
}
}
