% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(n, a, b, N, x1, y1)
}
\arguments{
\item{n}{the parameter of binomial}

\item{a}{shape1 of beta function}

\item{b}{shape2 of beta function}

\item{N}{the number of markov chain}

\item{x1}{the initial value of x}

\item{y1}{the initial value of y}
}
\value{
a markov chain
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- gibbsC(100,10)
par(mfrow=c(2,1));
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
}
}
