% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC_98}
\alias{gibbsC_98}
\title{A Gibbs sample using Rcpp}
\usage{
gibbsC_98(n, a, b, M)
}
\arguments{
\item{n}{Binomial distribution coefficient}

\item{a}{the Beta distribution with parameters shape1}

\item{b}{the Beta distribution with parameters shape2}

\item{M}{the Sample size}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sample using Rcpp
}
\examples{
\dontrun{
n <- 20
a <- 2
b <- 4
M <- 3500
randomnum.C <- gibbsC(n,a,b,M)
}
}
