% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constraints.R
\name{GIC}
\alias{GIC}
\title{GIC computation.}
\usage{
GIC(lambda, z, y, beta_lambda)
}
\arguments{
\item{lambda}{regularization parameter of L1 penalty}

\item{z}{n x p log design matrix with sample size n and dimensionality p}

\item{y}{n-vector of response}

\item{beta_lambda}{coeffcient value of regression under regularization parameter lambda}
}
\value{
a GIC value under the data given
}
\description{
GIC value computation, which is defined by (Lin, W., Shi, P., Feng, R., Li, H, 2014).
}
\examples{
\dontrun{
p=30
beta = c(c(1,-0.8,0.6,0,0,-1.5,-0.5,1.2),rep(0,p-8))
data1 = data_generate(n=50,p=30,beta = beta)
result = lrate_GIC(X=data1[[1]],y=data1[[2]])
learning_rate_GIC = result[1]
GIC_value = result[2]

#estimate beta
beta_est = estimate(X=data1[[1]],y=data1[[2]],learning_rate = unlist(learning_rate_GIC))

#compute estimation error.
pred_error(x=data1[[1]],y=data1[[2]],beta_hat = beta_est)

#compute estimation accuracy. 
est_acc(beta_est,beta)
}
}
