% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giccdmm.R
\name{giccdmm}
\alias{giccdmm}
\title{Variable selection in regression with compositional covariates}
\usage{
giccdmm(y, z, lambda, constr = TRUE, maxit = c(20, 50))
}
\arguments{
\item{y}{the response variables}

\item{z}{the log-ratio matrix}

\item{lambda}{a vector of candidate values of tuning parameters}

\item{constr}{Whether the model conforms to zero-sum constraint}

\item{maxit}{the max times of iterations}
}
\value{
a list contains coefficients, GIC, and tuning parameters
}
\description{
Variable selection of compositional data performed by lasso or a modified algorithm
}
\examples{
\dontrun{
res <- giccdmm(y, z, lambda, constr=TRUE, maxit=20)
}
}
