% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNAR.R
\name{GNAR}
\alias{GNAR}
\title{Grouped Network Autoregression using R.}
\usage{
GNAR(A, Z, data, begin, end, K, t)
}
\arguments{
\item{A}{The adjacency matrix of the data. (numeric)}

\item{Z}{The  node specific covariates Z should be a vector or matrix.(numeric)}

\item{data}{The data matrix. (numeric)}

\item{begin}{The begin day of training set. (numeric)}

\item{end}{The end day of training set. (numeric)}

\item{K}{the number of clusters. (numeric)}

\item{t}{the number of days to forecast. (numeric)}
}
\value{
the forecast and estimation also with the accuracy of the forecast
}
\description{
Grouped Network Autoregression  by two steps estimation using R.
}
\examples{
\dontrun{
data(A)
data(Z)
data(data)
begin=1; end=503
K=2; t=10
GNAR(A,Z,data,begin,end,K,t)
}
}
