% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{inter.points}
\alias{inter.points}
\title{Find the intersection points of a kind of curves}
\usage{
inter.points(k)
}
\arguments{
\item{k}{A numeric vector}
}
\value{
A matrix of size 2*length(k). The first row is 1:k
and the second row is the corresponding root value.
}
\description{
Find the intersection points of A(k) in (0,sqrt(k)) of the curves
S_{k-1}(a) = P(t(k-1)>sqrt{frac{a^2(k-1)}{k-a^2}}) and
S_k(a) = P(t(k)>sqrt{frac{a^2k}{k+1-a^2}}),
where t(k) is a Student t random variable with k degrees of freedom.
}
\examples{
\dontrun{
k = c(4:25,100,500,1000)
root <- inter.points(k)
print(root)
}
}
