% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA23204185.R
\name{IPW}
\alias{IPW}
\title{IPW algorithm}
\usage{
IPW(M, K)
}
\arguments{
\item{M}{missing matrices}

\item{K}{the rank of missing matrices}
}
\value{
the right singular vectors of truly M
}
\description{
The full name of the IPW algorithm is Inverse Probability Weighting algorithm, which aims to 
handle missing matrices in order to obtain their right singular vectors.
}
