% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{jack_correlation}
\alias{jack_correlation}
\title{Compute a jackknife estimate of the bias and the standard error
of the correlation statistic}
\usage{
jack_correlation(x, y)
}
\arguments{
\item{x}{a numeric vector of sample}

\item{y}{a numeric vector of another sample}
}
\value{
a vector of size 1*2 which contains the jackknife estimate of bias and
the jackknife estimate of the standard error of the correlation statistic.
}
\description{
Compute a jackknife estimate of the bias and the standard error
of the correlation statistic
}
\examples{
\dontrun{
data(law)
attach(law)
jack_est <- jack_correlation(LSAT,GPA)
round(jack_est,4)
}
}
