% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{jack_pca}
\alias{jack_pca}
\title{Obtain the jackknife estimates of bias and standard
error of the proportion of variance explained by the first principal component.}
\usage{
jack_pca(X)
}
\arguments{
\item{X}{a data matrix}
}
\value{
a numeric vector contains the jackknife estimate of bias and se.
}
\description{
Obtain the jackknife estimates of bias and standard
error of the proportion of variance explained by the first principal component.
}
\examples{
\dontrun{
data(scor)
ja <- jack_pca(scor)
print(round(ja,4))
}
}
