% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirs.R
\name{kernal}
\alias{kernal}
\alias{kernel}
\title{A kernel function of naive density estimator}
\usage{
kernel(x, Xi, h)
}
\arguments{
\item{x}{independent variable}

\item{Xi}{the sample}

\item{h}{bandwidth}
}
\value{
the value of the kernel function in x.
}
\description{
the value of naive density estimator kernel with sample X_i.
}
\examples{
\dontrun{
n=1000
x1=rnorm(n,0,1)
x2=rnorm(n,1,0.3)
index<-sample(c(1,0),n,replace = T,prob = c(0.3,0.7))
X=index*x1+(1-index)*x2
h=0.05
x=seq(-2,3,0.01)
y=c(rep(0,length(x)))
for(i in 1:n){
y=y+kernel(x,X[i],h)/(n*h)
 }
plot(x,y,type='l',main ='h=0.05')
}
}
