% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{knn.est}
\alias{knn.est}
\title{the kth nearest neighbor density estimate function.}
\usage{
knn.est(x, k, xrange, yrange)
}
\arguments{
\item{x}{an input data matrix}

\item{k}{the maximum number of nearest neighbors to search.}

\item{xrange}{the range of x sample}

\item{yrange}{the range of y sample}
}
\value{
a matrix contains the value of estimation
}
\description{
Use function knn.dist() in package FNN
to write a kth nearest neighbor density estimate function.
}
\examples{
\dontrun{
data(faithful)
X <- as.matrix(faithful)
xrange <- seq(from=1,to=6,by=.1)
yrange <- seq(from=40,to=100,by=.5)
k <- 5
fit_knnde <- knn.est(X, k, xrange, yrange)
persp(xrange, yrange, fit_knnde, phi = 30, theta = 45, border = 0, col = "blue")
}
}
