% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{knn.im}
\alias{knn.im}
\title{knn imputation}
\usage{
knn.im(data, k)
}
\arguments{
\item{data}{the data include missing value}

\item{k}{the number of neighbor needed to taken into account}
}
\value{
a data after imputation
}
\description{
use the method of knn to impute missing value.
}
\examples{
\dontrun{
library(VIM)
library(DMwR)
knn.im(sleep,k=10)
} 
}
