% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/functionsH.R
\name{lap}
\alias{lap}
\title{Laplace density function}
\usage{
lap(x)

lap(x)
}
\arguments{
\item{x}{a real number}
}
\value{
the density at number x

standard Laplace distribution density at x
}
\description{
Standard Laplace distribution density  using Rcpp

This function is defined to make it easier to write
}
\examples{
\dontrun{
lap(0)
}
\dontrun{
lap(0)
lap(0.5)
}
}
