% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lars.R
\name{Lars}
\alias{Lars}
\title{Lars algorithm for linear regression}
\usage{
Lars(y, X)
}
\arguments{
\item{y}{a vector of response variable}

\item{X}{design matrix}
}
\value{
a list including estimated coefficients and Cps
}
\description{
Lars algorithm for linear regression
}
\details{
This is some descriptio of this function.
}
\examples{
\dontrun{
y <- rnorm(100)
X <- matrix(runif(500), nrow = 100)
Lars(y, X)
} 

}
