% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lars.R
\name{Lars_update}
\alias{Lars_update}
\title{update step for Lars algorithm for linear regression}
\usage{
Lars_update(y, X, bt)
}
\arguments{
\item{y}{a vector of response variable}

\item{X}{design matrix}

\item{bt}{current coefficient vector}
}
\value{
coefficient vector after update
}
\description{
update step for Lars algorithm for linear regression
}
\details{
This is some descriptio of this function.
}
\examples{
\dontrun{
y < rnorm(100)
X < matrix(runif(500), nrow = 100)
bt < rnorm(5)
Lars_update(y, X, bt)
} 
}
