% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locallikelihood.R
\name{LCV}
\alias{LCV}
\title{Cross-validation}
\usage{
LCV(X, Y, h, n)
}
\arguments{
\item{X}{the sample vector X (numeric)}

\item{Y}{the sample vector Y (numeric)}

\item{h}{the bandwidth of the kernel function (numeric)}

\item{n}{the size of the sample}
}
\value{
the the optimal bandwidth of the kernel function
}
\description{
Use the cross-validation to choose the optimal bandwidth of the kernel function in the exponential generalized linear model
}
\examples{
\dontrun{
    n <- 200
    truebeta <- c(2,2)
    lambda <- function(x) exp(truebeta[1] + truebeta[2] * x)
    X <- runif(n = n, -3, 3)
    Y <- rexp(n = n, rate = lambda(X))
    h <- seq(0.1,1, by = .1)
    LCV(X,Y,h,n)
}
}
