% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{leave2out_cv}
\alias{leave2out_cv}
\title{leave_two_out cross validation}
\usage{
leave2out_cv(X, Y)
}
\arguments{
\item{X}{a data sample which represents the response varirable.}

\item{Y}{another data sample which represents the predictive variable.}
}
\value{
a numeric vector of residual sum of squares of all models.
}
\description{
Use leave-two-out cross validation to compare the models.
}
\examples{
\dontrun{
data(ironslag)
attach(ironslag)
cv <- leave2out_cv(chemical,magnetic)
print(cv)
}
}
