% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{leftx}
\alias{leftx}
\title{SFldichotomy}
\usage{
leftx(lam, alpha, prec)
}
\arguments{
\item{lam}{function value}

\item{alpha}{the shape parameter of t he gamma function}

\item{prec}{balabal}
}
\description{
Find the solution of the function by dichotomy
}
