% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locallikelihood.R
\name{likelifunc}
\alias{likelifunc}
\title{Estimate of the conditional expectation}
\usage{
likelifunc(x, X, Y, h, n)
}
\arguments{
\item{x}{evaluation grid}

\item{X}{the sample vector X (numeric)}

\item{Y}{the sample vector Y (numeric)}

\item{h}{the bandwidth of the kernel function (numeric)}

\item{n}{the size of the sample}
}
\value{
the value of the (i) item of the Log-likelihood of beta
}
\description{
Estimate of the conditional expectation in the exponential generalized linear model, Compare the true value with the estimate
}
\examples{
\dontrun{
    n <- 200
    truebeta <- c(2,2)
    lambda <- function(x) exp(truebeta[1] + truebeta[2] * x)
    X <- runif(n = n, -3, 3)
    Y <- rexp(n = n, rate = lambda(X))
    h <- seq(0.1,1, by = .1)
    x <- seq(-3, 3, l = 501)
    likelifunc(x,X,Y,h,n)
}
}
