% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/limit_curves_entropy_complexity.R
\name{limit_curves}
\alias{limit_curves}
\title{Limit curves in the Entropy-Complexity plane}
\usage{
limit_curves(ndemb, fun = "min")
}
\arguments{
\item{ndemb}{Embedding dimension}

\item{fun}{Whether the upper (max) or lower (min) limit curve should be computed}
}
\value{
A list with two entries
}
\description{
Compute the limit curves in the Entropy Complexity plane
}
\details{
This function returns the respective limit curve.
}
\author{
Sebastian Sippel
}
\references{
none
}

