% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{loop_lapply}
\alias{loop_lapply}
\title{Use both for loops and \code{lapply()} to fit linear models.}
\usage{
loop_lapply(formulas, data)
}
\arguments{
\item{formulas}{a list of models}

\item{data}{from datasets}
}
\value{
A list composed of two lists of fitting models
}
\description{
Use both for loops and \code{lapply()} to fit linear models in data mtcars.
}
\examples{
\dontrun{
data(mtcars)
attach(mtcars)
formulas = list(
 mpg ~ disp,
 mpg ~ I(1 / disp),
 mpg ~ disp + wt,
 mpg ~ I(1 / disp) + wt
 )
ll <- loop_lapply(formulas,mtcars)
print(ll$lo)
print(ll$la)
}
}
