% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIVE-R.R
\name{lplqnorm}
\alias{lplqnorm}
\title{A function computing $l_p/l_q$ norm of a matrix}
\usage{
lplqnorm(M, p, q)
}
\arguments{
\item{M}{the matrix to be calculated}

\item{p}{the parameter of the norm}

\item{q}{the parameter of the norm}
}
\value{
the $l_p/l_q$ norm of a matrix
}
\description{
A function computing $l_p/l_q$ norm of a matrix
}
\examples{
\dontrun{
    A <- matrix(11:16, nrow=3, ncol=2)
    lplqnorm(A, 1, 2)
}
}
