% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_class.R
\name{make_p3_class}
\alias{make_p3_class}
\title{Create new class p3_class}
\usage{
make_p3_class(x)
}
\arguments{
\item{x}{numeric vector of length N}
}
\value{
nothing, unless print function called
}
\description{
Create new class p3_class
}
\details{
This constructor function creates a new S3 class called p3_class and the print method prints the number of observations of the class.
}
\examples{
set.seed(4)
x <- rnorm(1:10)
make_p3_class(x)
print(x)
}
