% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Mardia}
\alias{Mardia}
\title{Mardia's multivariate skewness test.}
\usage{
Mardia(mydata)
}
\arguments{
\item{mydata}{a numeric matrix of data}
}
\value{
0 or 1, 1 represents the test being rejected, which 0 represents the test cannot be rejected.
}
\description{
Mardia's multivariate skewness test.
}
\examples{
\dontrun{
mu <- c(0,0,0)
sigma <- matrix(c(1,0,0,0,1,0,0,0,1),nrow=3,ncol=3)
m=1000
n<-c(10, 20, 30, 50, 100, 500)
a=numeric(length(n))
for(i in 1:length(n)){
  a[i]=mean(replicate(m, expr={
    mydata <- mvrnorm(n[i],mu,sigma)
    Mardia(mydata)
  }))
}
}
}
