% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{maxd4}
\alias{maxd4}
\title{Maximum curve of time-causal entropy-complexity plane at ndemb=4}
\format{A data frame with 2139 rows and 2 columns:
\describe{
  \item{x}{x-values of minimum curve if ndemb==4}
  \item{y}{y-values of minimum curve if ndemb==4}
  ...
}}
\source{
Computed based on Martin, M. T., A. Plastino, and O. A. Rosso. "Generalized statistical complexity measures: Geometrical and analytical properties." Physica A: Statistical Mechanics and its Applications 369.2 (2006): 439-462.
}
\usage{
maxd4
}
\description{
Maximum curve of time-causal entropy-complexity plane at ndemb=4
}
\keyword{datasets}

