% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{maxout}
\alias{maxout}
\title{Find the maximum number of extreme observations}
\usage{
maxout(x, y)
}
\arguments{
\item{x}{the one sample vector}

\item{y}{the other sample vector}
}
\value{
the maximum number of extreme observations from two samples
}
\description{
Count the number of extreme observations in each sample that 
are not in the other sample. Calculate the number of maximum extreme points.
}
\examples{
\dontrun{
x<-1:10
y<-3:12
maxout(x,y)
}
}
