% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCgnr.R
\name{MCgnr}
\alias{MCgnr}
\title{a random walk Metropolis sampler}
\usage{
MCgnr(m, x0, sd)
}
\arguments{
\item{m}{the length of chain}

\item{x0}{the initial value given to the sampler}

\item{sd}{the standard deviation of the proposal normal distribution}
}
\value{
a random sample of size \code{m}
}
\description{
generate the standard Laplace distribution
}
\examples{
\dontrun{
x <- MCgnr(1e4, 0, 1)
par(mfrow=c(1,1));
plot(x,type='l')
}
}
