% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCtestforAR.R
\name{MCtestforAR}
\alias{MCtestforAR}
\alias{MCtestar}
\title{MC test for AR(1)}
\usage{
MCtestar(data, phi, M, H, N)
}
\arguments{
\item{data}{data set}

\item{phi}{the unit roots fot test}

\item{M}{length of chain of phi of stage 2}

\item{H}{length of chain of phi of stage 1}

\item{N}{number of reputation for the test}
}
\value{
power of value
}
\description{
Monte Carlo test for autogressive panel
}
\examples{
\dontrun{
data=ar(0.6,10,rep(0,5))
resault=MCtestar(data,0.6,20,20,100)
print(resault)
}
}
