% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp001.R
\name{MDS}
\alias{MDS}
\title{Multi-Dimensional Scaling using R}
\usage{
MDS(X, n)
}
\arguments{
\item{X}{a matrix composed of m d-dimensional vectors, the number of rows is m}

\item{n}{Low-dimensional space dimension}
}
\value{
z The representation of the original vector in the low-dimensional space, each row is the low-dimensional coordinates of a sample \code{n}
}
\description{
Realize multi-dimensional scale transformation with R function
}
\examples{
\dontrun{
X <- matrix(c(3,5,2,3,4,8,9,6,1,7,5,9,8,6,4,4),nrow=4)
MDS(X, n=2)
}
}
