% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Metropolis}
\alias{Metropolis}
\title{A Metropolis sampler using Rcpp}
\usage{
Metropolis(x0, sigma, N)
}
\arguments{
\item{x0}{the initial point of the series}

\item{sigma}{the standard deviation}

\item{N}{the sample size}
}
\value{
a random sample of size \code{N}
}
\description{
A Metropolis sampler using Rcpp
}
\examples{
\dontrun{
x0 = 0;
sigma = 2;
N = 1000;
x = Metropolis(x0, sigma, N)
}

}
