% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHchain.R
\name{MHT}
\alias{MHT}
\alias{MHC}
\title{Generate a Metropolis-Hastings chain with student t(df, mu) distribution or Cauchy(l, s) distribution with proposal distribution Normal(xt, sigma).}
\usage{
MHT(n, df, ncp = NULL, a = 1, sigma = 1)

MHC(n, l, s, a = 1, sigma = 1)
}
\arguments{
\item{n}{length of chain. (better if not so small).}

\item{df}{degrees of freedom (maybe non-integer) of t distribution.}

\item{ncp}{non-centrality parameter. If omitted, use the central t distribution.}

\item{a}{initial value of the chain. 1 is default.}

\item{sigma}{sd of proposal distribution. 1 is default.}

\item{l}{location of Cauchy distribution.}

\item{s}{scale of Cauchy distribution.}
}
\value{
return a numerical vector (chain) with length n.

a vector (chain) of Cauchy distribution.
}
\description{
The Normal(xt, sigma) (random walk Metropolis sampler) is an example of a Metropolis sampler. \code{MHT} generates
a chain with asymptotic student t(df, mu) distribution while \code{MHC} generates Cauchy(l, s) distribution.
}
\section{Functions}{
\itemize{
\item \code{MHC}: replace t distribution with Cauchy distribution
}}

\examples{

\dontrun{
## a M-H chain with central t(3) distribution with length 1000
MHT(1000, 3, sigma = 2)

## a M-H chain with Cauchy(0, 1), stored in x.
n = 2000
x = MHC(n, 0, 1, a = -5)

## and you can see how the chain converge.
plot(x, type = "l", xlab = "x", xlim = 1:n, ylim = range(x))
}

}
