% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{mind5}
\alias{mind5}
\title{Minimum curve of time-causal entropy-complexity plane at ndemb=5}
\format{A data frame with 500 rows and 2 columns:
\describe{
  \item{x}{x-values of minimum curve if ndemb==5}
  \item{y}{y-values of minimum curve if ndemb==5}
  ...
}}
\source{
Computed based on Martin, M. T., A. Plastino, and O. A. Rosso. "Generalized statistical complexity measures: Geometrical and analytical properties." Physica A: Statistical Mechanics and its Applications 369.2 (2006): 439-462.
}
\usage{
mind5
}
\description{
Minimum curve of time-causal entropy-complexity plane at ndemb=5
}
\keyword{datasets}

