% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{mind6}
\alias{mind6}
\title{Minimum curve of time-causal entropy-complexity plane at ndemb=6}
\format{A data frame with 500 rows and 2 columns:
\describe{
  \item{x}{x-values of minimum curve if ndemb==6}
  \item{y}{y-values of minimum curve if ndemb==6}
  ...
}}
\source{
Computed based on Martin, M.T., Plastino, A. and Rosso, O.A., 2006. Generalized statistical complexity measures: Geometrical and analytical properties. Physica A: Statistical Mechanics and its Applications, 369(2), pp.439-462.
}
\usage{
mind6
}
\description{
Minimum curve of time-causal entropy-complexity plane at ndemb=6
}
\keyword{datasets}

