% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{MissingValue}
\alias{MissingValue}
\title{Missing value judgment}
\usage{
MissingValue(x)
}
\arguments{
\item{x}{data in the form of a matrix or list or data frame}
}
\value{
a logical value: 0 means there is no missing value; 1 mean there is missing value.
}
\description{
determine whether the data has missing values (the missing value form is NA)
}
