% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{mixnorm}
\alias{mixnorm}
\title{Generate random data from a mixnormal distribution}
\usage{
mixnorm(n)
}
\arguments{
\item{n}{a scalar which represents the amount of generated data points.}
}
\value{
the generated data
}
\description{
Generate n data points from the distribution
0.3 N(0, 1) + 0.7 N(1, 0.32), use the bandwidth selection methods
in R package kedd, and draw the estimated density curves with
different bandwidths in one plot.
}
\examples{
\dontrun{
n <- 100
set.seed(0)
mixnorm(n)
}
}
