% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{mtcars}
\alias{mtcars}
\title{data}
\description{
A dataset used to illustrate the performance of loops and \code{lapply()}.
}
\examples{
\dontrun{
data(mtcars)
attach(mtcars)
formulas = list(
 mpg ~ disp,
 mpg ~ I(1 / disp),
 mpg ~ disp + wt,
 mpg ~ I(1 / disp) + wt
 )
ll <- loop_lapply(formulas,mtcars)
print(ll$lo)
print(ll$la)
}
}
