% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMmixnormal.R
\name{myEM}
\alias{myEM}
\title{EM algorithm for Gaussian Mixtures with 2 components}
\usage{
myEM(M)
}
\arguments{
\item{M}{the model of initial Gaussian Mixtures}
}
\value{
a list containing at least the following components: 

\code{data} samples to induct EM algorithm on 

\code{lambda} weights for the two normal distributions 

\code{mu} mean of normal distributions 

\code{sigma} standard deviation of normal distributions
}
\description{
estimate parameters for Gaussian Mixtures
}
\examples{
\dontrun{
data(faithful)
n <- nrow(faithful)
mu0 <- list(c(2,60), c(4.5,80))
s0 <- list(matrix(c(1,0,0,1),nrow=2), matrix(c(1,0,0,1),nrow=2))
M0 <- list(data=faithful, lambda=c(0.5, 0.5), mu=mu0, sigma=s0)
M <- myEM(M0) 
}
}
