% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mygibbs}
\alias{Mygibbs}
\title{A Gibbs sampler using Rcpp}
\usage{
Mygibbs(x, y, N)
}
\arguments{
\item{x}{the origin value of the random variable X}

\item{y}{the origin value of the random variable Y}

\item{N}{the number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- Mygibbs(1,0.1,1000)
par(mfrow=c(2,1));
plot(rnC[1,],type='l')
plot(rnC[2,],type='l')
}
}
