% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans.R
\name{MyKmeans}
\alias{MyKmeans}
\title{K-means function}
\usage{
MyKmeans(data, k)
}
\arguments{
\item{data}{the data that you want to classify (numeric)}

\item{k}{The number of categories you want to classify (numeric)}
}
\value{
the value of the (i) item of the Log-likelihood of beta
}
\description{
K-means: a kind of clustering algorithm; unsupervised learning.
}
\examples{
\dontrun{
 MyKmeans(swiss,3)
}
}
