% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwMetropolis.R
\name{mylapply}
\alias{mylapply}
\title{mylapply}
\usage{
mylapply(X, FUN, FUN.VALUE, simplify = FALSE)
}
\arguments{
\item{X}{data}

\item{FUN}{function}

\item{FUN.VALUE}{FUN.VALUE in vapply}

\item{simplify}{turn out into array}
}
\value{
function
}
\description{
Implement a combination of Map() and vapply() to create an lapply() variant that iterates in parallel over all of its inputs and stores its outputs in a vector (or a matrix).
}
