% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{naive}
\alias{naive}
\title{naive}
\usage{
naive(data)
}
\arguments{
\item{data}{the data include missing value}
}
\value{
a data after delete
}
\description{
delete data with missing values
}
\examples{
\dontrun{
library(VIM)
naive(sleep)
} 
}
