% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NearestNUserNeighborhood.R
\name{NearestNUserNeighborhood}
\alias{NearestNUserNeighborhood}
\title{The Nearest N User Neighborhood}
\usage{
NearestNUserNeighborhood(S, n)
}
\arguments{
\item{S}{S is the similarity matrix of users}

\item{n}{n is the number of neighborhood.}
}
\value{
The Nearest N User Neighborhood
}
\description{
The function computes the nearest n user neighborhood
}
\examples{
\dontrun{
data(S)
NEIGHBORHOOD_NUM <- 2
N <- NearestNUserNeighborhood(S, NEIGHBORHOOD_NUM)
N
}
}
