% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp001.R
\name{para_EM}
\alias{para_EM}
\title{EM algorithm estimates Gaussian mixture model parameters}
\usage{
para_EM(x, mean, sd = NULL)
}
\arguments{
\item{x}{Sample random number}

\item{mean}{Mixed distribution mean vector}

\item{sd}{Mixed distribution sd vector}
}
\value{
Mean estimation, standard deviation estimation, mixed component probability estimation \code{n}
}
\description{
Use R language to estimate Gaussian mixture model parameters through EM algorithm
}
\examples{
\dontrun{
n <- 1000
mean_s <- c(1, 7)
y <- sample(c("head", "tail"), size = n, replace = TRUE, prob = c(0.25, 0.75))
x <- rnorm(n = 1000, mean = mean_s[1])
tails <- y \%in\% c("tail")
x[tails] <- rnorm(sum(tails), mean = mean_s[2])
para_EM(x, mean = c(0, 1), sd = c(1, 1))
}
}
