% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisrand.R
\name{poisrand}
\alias{poisrand}
\title{Simulation Poisson Process}
\usage{
poisrand(n, lambda)
}
\arguments{
\item{n}{the number of samples}

\item{lambda}{the mean of the Poisson distribution}
}
\value{
a random sample of size \code{n} for Poisson Process
}
\description{
A Poisson Process sampler using R
}
\examples{
temp=poisrand(1000,5)
head(temp)
mean(temp)
}
