% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{power_stu}
\alias{power_stu}
\title{Power of the skewness test of student t distributions.}
\usage{
power_stu(para, level = 0.05, n = 50, m = 2500)
}
\arguments{
\item{para}{parameter vector of t distributions}

\item{level}{significance level}

\item{n}{size of sample drawn from t distribution for each experiment}

\item{m}{number of repeated experiments}
}
\value{
the power of skewness test
}
\description{
Estimate the power of the skewness test of normality against 
student t(nu) distributions by using the sample skewness 
statistic function sk(). The value of power is the proportion of null hypothesis 
(skewness of distribution is 0) rejected in m replicates.
}
\examples{
\dontrun{
alpha<-c(seq(0.5,10,0.5),11:30)
pwr<-power_stu(alpha)
plot(para1, pwr, type = "b",xlab = bquote(alpha), ylim = c(0,0.1))
}
}
