% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/powernoise.R
\name{powernoise}
\alias{powernoise}
\title{A function to generate k-noise}
\usage{
powernoise(k, N)
}
\arguments{
\item{k}{Power law scaling exponent}

\item{N}{number of samples to generate}
}
\value{
A named list with three entries is returned.
x - N x 1 vector of power law samples
}
\description{
Generates samples of power law noise.
}
\details{
Generates samples of power law noise.
The power spectrum of the signal scales as f^(-k). The R function uses fft(),
 similarly to the knoise_fft Matlab function.
}
\examples{
powernoise_series = powernoise(k=2, N=10000)
}
\author{
Sebastian Sippel and Holger Lange
}

