% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.p3_class}
\alias{print.p3_class}
\title{Print p3_class}
\usage{
\method{print}{p3_class}(x)
}
\arguments{
\item{x}{list of class p3_class}
}
\value{
print x
}
\description{
Print p3_class
}
\examples{
set.seed(4)
x <- rnorm(1:10)
class(x) <- "p3_class"
print.p3_class(x)
}
