% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{pwr_F}
\alias{pwr_F}
\title{Compute the F test of equal variance.}
\usage{
pwr_F(n, alpha.hat)
}
\arguments{
\item{n}{a numeric vector}

\item{alpha.hat}{the confidence level}
}
\value{
a dataframe contains n and power of Count 5 test and F test.
}
\description{
Compute the F test of equal variance at significance level 0.055 and compare the power of the Count Five test and F test for small, medium, and large sample sizes.
}
\examples{
\dontrun{
set.seed(1027)
alpha.hat <- 0.055
n <- c(10, 20, 50, 100, 500, 1000)
pwr <- pwr_F(n,alpha.hat)
print(pwr)
}
}
