% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{pwr_t}
\alias{pwr_t}
\title{Estimate the power of the skewness test of
normality against t(v) distributions}
\usage{
pwr_t(v)
}
\arguments{
\item{v}{the parametric of t(v) distributions.}
}
\value{
a list contains the power and its se
}
\description{
Estimate the power of the skewness test of
normality against t(v) distributions
}
\examples{
\dontrun{
v <- seq(1,20)
pwr <- pwr_t(v)$pwr
se <- pwr_t(v)$se
plot(v,pwr,type="b",xlab="v",ylab="pwr",ylim=c(0,1),pch=16)
abline(h=0.1,lty=2)
lines(v, pwr+se, lty = 4)
lines(v, pwr-se, lty = 4)
}
}
