% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/chaotic_maps.R
\name{quadratic_map}
\alias{quadratic_map}
\title{A function to generate a time series from the Quadratic map}
\usage{
quadratic_map(N, k, start="rand", disregard_N=0)
}
\arguments{
\item{N}{length of the time series that is to be generated}

\item{k}{Quadratic map parameter}

\item{start}{start value. Default is to random.}

\item{disregard_N}{Number of values at the beginning of the series to disregard}
}
\value{
A vector of length N
}
\description{
Generates a time series from the Quadradtic map
}
\examples{
quadratic_map(N = 10^4, k=1.4)
}
\author{
Sebastian Sippel
}
\references{
Grebogi, C., Ott, E. and Yorke, J.A., 1983. Crises, sudden changes in chaotic attractors, and transient chaos. Physica D: Nonlinear Phenomena, 7(1-3), pp.181-200.
}

