% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework4.R
\name{question2_1c}
\alias{question2_1c}
\title{question2.1_in_HW3}
\usage{
question2_1c(m, n, k)
}
\description{
The first function used of question2 in HW3
}
\examples{
\dontrun{
q2_1<-question2_1c(10000,10,2)
}
}
