% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework4.R
\name{question2_2c}
\alias{question2_2c}
\title{question2.2_in_HW3}
\usage{
question2_2c(m, x)
}
\description{
The second function used of question2 in HW3
}
\examples{
\dontrun{
q2_2<-question2_2c(10000,1.95)
}
}
