% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework2.R
\name{question4a}
\alias{question4a}
\title{question4_in_HW1}
\usage{
question4a(N)
}
\description{
The function used of question4 in HW1
}
\examples{
\dontrun{
tm1<-microbenchmark::microbenchmark(
  q4<-question4a(1000)
)
}
}
