% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/transformPermCoding.R
\name{rank_to_permutation}
\alias{rank_to_permutation}
\title{A function to convert a "ranks-based" permutation notation to an "index-based" permutation scheme.}
\usage{
rank_to_permutation(pattern, permutation.notation)
}
\arguments{
\item{pattern}{A numeric vector that denotes a permutation pattern.}

\item{permutation.notation}{The permutation notation that should be used. Could be "Olivares.2012" or "Keller.2005".}
}
\value{
A numeric vector, which contains the transformed permutation.
}
\description{
Converts permutations denoted by ranks to permutations denoted by indices and back.
}
\details{
This function converts ranks to indices and back.
}
\author{
Sebastian Sippel
}
\references{
Sebastian Sippel (2014). Master Thesis. University of Bayreuth.
}

