% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rBC.R
\name{rBC}
\alias{rBC}
\title{Generate data with Beta(a, b) distribution or Cauchy(l, s) distribution with Acceptance-rejection method.}
\usage{
rBC(n, a, b, type = NULL, p.value = FALSE)
}
\arguments{
\item{n}{the size of sample.}

\item{a}{the first argument of Beta distribution or the location of Cauchy distribution.}

\item{b}{the second argument of Beta distribution or the scale of Cauchy distribution.}

\item{type}{the type of distribution, \emph{Beta} or \emph{Cauchy}.}

\item{p.value}{logical; FALSE is default. If TRUE, show the p-value of Cramer-von Mises test between samples generated by \code{rBC} and \code{rbeta} or \code{rcauchy}.}
}
\value{
a numerical vector and p-value if TRUE.
}
\description{
\code{rBC} use Acceptance-rejection method to generate data and then can use \code{CVM.test} to test data with the sample generated
by \code{rbeta} and \code{rcauchy}.
}
\examples{

\dontrun{
## a sample with Beta(2, 3) and no p-value.
rBC(2, 3, type = "Beta")

## a sample with Cauchy(0, 1) and return p-value at the same time.
rBC(0, 1, type = "Cauchy", p = TRUE)
}

}
\seealso{
\code{\link{CVM.test}}.
}
