% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{rightx}
\alias{rightx}
\title{SFrdichotomy}
\usage{
rightx(lam, alpha, prec)
}
\arguments{
\item{lam}{function value}

\item{alpha}{the shape parameter of t he gamma function}

\item{prec}{balala}
}
\description{
Find the solution of the function by dichotomy
}
