% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwMetropolis.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{rwMetropolis}
\usage{
rw.Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{variance of normal distrubtion}

\item{x0}{Initial value}

\item{N}{length of chain}
}
\value{
dataframe include number of candidate points rejected and chain
}
\description{
Implement a random walk Metropolis sampler for generating the standard Laplace distribution
}
\examples{
\dontrun{
N = 2000
sigma = c(.05, .5, 2, 16)
x0 = 25
rw1 = rw.Metropolis(sigma[1],x0,N)
}
}
