% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMetropolisC}
\alias{rwMetropolisC}
\title{A random walk Metropolis sampler by using Rcpp}
\usage{
rwMetropolisC(N, sigma, x0)
}
\arguments{
\item{N}{the number of samples}

\item{sigma}{the standard deviation}

\item{x0}{the initial value}
}
\value{
a list contains the generated samplers x of size n and the counter k which records the number of rejected candidate points
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution by using Rcpp
}
\examples{
\dontrun{
N <- 2000; sigma <- 2; x0 <- 25
rw <- rwMetropolisC(N,sigma,x0)
plot(x=1:N,y=rw$x,type="l")
}
}
