\name{SA23204185-package}
\alias{SA23204185-package}
\docType{package}
\title{
  IPW and Distribution
}
\description{
  Reproduction of the IPW and distributed algorithms as well as a calculator of a delivery problem
}
\details{
  Develop an R package to reproduce the algorithms in the paper 'Distributed learning in recommender systems,' including the IPW algorithm for recovering the right singular values of the missing matrix and the Distribution algorithm for distributed processing.
This package also contains a cpp function targetoutofn  to calculate the  probability of  a certain number 'target' out of n (target <= n) delivery man with respectively different probability to accept an order.
}
\author{
Zhou Runlin

Maintainer: Zhou Runlin <zrl1644@mail.ustc.edu.cn>
}
\keyword{ package }
