% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{sapply_t_p}
\alias{sapply_t_p}
\title{Use \code{sapply()} to extract the p value from t-test}
\usage{
sapply_t_p()
}
\value{
the p values of every trial
}
\description{
The fucntion simulates the performance of a t-test
 for non-normal data and use \code{sapply()} and an anonymous
  function to extract the p-value from every trial.
}
\examples{
\dontrun{
set.seed(123)
p <- sapply_t_p()
print(p)
}
}
