% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi_Li.R
\name{semi_Li}
\alias{semi_Li}
\title{Semiparametric partially linear model with Li method}
\usage{
semi_Li(y, X, Z, type = 1, h, verbose = FALSE)
}
\arguments{
\item{y}{the response vector}

\item{X}{the data of parametric part}

\item{Z}{the data of non-parametric part}

\item{type}{the type of kernal function}

\item{h}{the bandwidth of kernal function}

\item{verbose}{the type of kernal function}
}
\value{
a list of the regression parameters
}
\description{
Use R to do  Semiparametric partially linear model with Li method
}
\examples{
\dontrun{
semi_Robinson(y,X,Z,h=1)
} 
}
