% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SGD_C}
\alias{SGD_C}
\title{A realization of stochastic gradient descent(SGD) algorithm using Rcpp}
\usage{
SGD_C(matrix, result, theta, loss)
}
\arguments{
\item{matrix}{the features of training data}

\item{result}{the estimation of the interesting feature}

\item{theta}{initialization of paramaters}

\item{loss}{initialization of flag variable to judge the convergency}
}
\value{
the optimized paramaters \code{n}
}
\description{
A SGD realization using Rcpp
}
\examples{
\dontrun{
matrix <- matrix(c(1,4,2,5,5,1,4,2), nrow = 4, ncol = 2)
result <- c(19,26,19,20)
theta <- c(2,5)
loss <- 1000.0
res <- SGD_C(matrix, result, theta, loss)
print(res)
}
}
