% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sirscore}
\alias{sirscore}
\title{sirscore}
\usage{
sirscore(
  A,
  y,
  a,
  x0,
  maxsize,
  delta = 1e-06,
  thresh = 1e-06,
  maxiter = 50L,
  tol = 1e-06
)
}
\arguments{
\item{A}{a matrix with dim n*p}

\item{y}{a vector with length n}

\item{a}{shape parameter of SICA penalty controling its maximum concavity}

\item{x0}{initial set}

\item{maxsize}{maximum size of the sparse model}

\item{delta}{param of SIRS}

\item{thresh}{threshold of the circulation. Default = 1e-6}

\item{maxiter}{maximum number of iterations. Default = 50}

\item{tol}{tolerance for updating.  Default = 1e-6}
}
\value{
a coeffcient value in sparse recovery problem
}
\description{
step 1 implemetation of algorithm sequentially and iteratively reweighted squares(SIRS).
}
