% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{sk}
\alias{sk}
\title{Compute the sample skewness statistic.}
\usage{
sk(x)
}
\arguments{
\item{x}{the sample vector}
}
\value{
sample skewness statistic
}
\description{
The calculated statistic is used for skewness test of normality 
and has an asymptotic distribution of N(0,n/6).
}
\examples{
\dontrun{
x<-rnorm(100)
sktest<-(abs(sk(x))>= qnorm(.975, 0, sqrt(6/n)))
}
}
