% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwMetropolis.R
\name{solve}
\alias{solve}
\title{get roots of 11.4}
\usage{
solve(k)
}
\arguments{
\item{k}{degree of freedom}
}
\value{
root
}
\description{
for k = 4 : 25, 100, 500, 1000, where t(k)  is a Student t random variable with k degrees of freedom.
}
