\name{StatComp18039-package}
\alias{StatComp18039-package}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
  Several R function are provided in this package (\code{dRayleigh} and \code{pRayleigh}  \code{qRayleigh} and \code{rRayleigh} and \code{figR} and \code{figC} and \code{cauchy.chain})
}
\author{
18039

Maintainer: Jie Hu <hujie@mail.ustc.edu.cn>
}
\keyword{ package }


