\name{StatComp20020}
\alias{StatComp20020}
\docType{package}
\title{
  Final Homework for Statistical Computing
}
\description{
  Two functions to sovle CURE and SeCURE (with several R and Rcpp functions for homeworks). 
}
\author{
Letian Li

Maintainer: Letian Li <letianli1998@mail.ustc.edu.cn>
}
\keyword{ package }
