\name{StatComp20050-package}
\alias{StatComp20050-package}
\alias{StatComp20050}
\docType{package}
\title{
  this function aim to find subgroup from a data with dropout.
}
\description{
  find the intercept of each individual,then cluster them according to different intercepts.
}
\details{
  The respond variable should with the kind of: row belong to a individual.
  The independent variable with the form of: have p rows and m*n columns.
}
\author{
Huicong Yu

Maintainer: Huicong Yu <190861883@qq.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
