% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{StatComp20070-package}
\alias{StatComp20070-package}
\alias{StatComp20070}
\title{StatComp20070.}
\description{
StatComp20070 package.
}
\details{
a visualized package for Statcomp20070
}
\keyword{package}
