\name{StatComp20082-package}
\alias{StatComp20082-package}
\docType{package}
\title{
  This package is the final project for 'Statistical Computing' course, containing two R functions and my homework summary.
}
\description{
  A function of Metropolis sampler to generate the standard Laplace distribution (\code{MCgnr})
  
  A function of EM algorithm to esitimate the parameters of Gaussian Mixtures with 2 components (\code{myEM})  
}
\details{
  Some functions and homework summary.
}
\author{
Xiaohui Yin

Maintainer: Xiaohui Yin <xiaohuiyyy@gmail.com>
}
\keyword{ package }
