\name{StatComp21001-package}
\alias{StatComp21001-package}
\alias{StatComp21001}
\docType{package}
\title{
  Multi-Dimensional Scaling and the estimators of Gaussian mixture model parameters
}
\description{
  using R to estimate Gaussian mixture model parameters and Multi-Dimensional Scaling.
}
\details{
Lu Qi

Maintainer: Lu Qi <hfut2016214493@163.com>
}
\author{
Your Name, email optional.

Maintainer: Your Name <your@email.com>
}
\keyword{ package }
