\name{StatComp21020-package}
\alias{StatComp21020-package}
\alias{StatComp21020}
\docType{package}
\title{
  Course assignment for 'Statistical Computing'.
}
\description{
  This package includes a Rcpp function, several functions to solve the  local likelihood estimation for  exponential generalized linear model and a funtions to apply the K-means algorithm.
}
\details{
1.a Rcpp function: the application of the Gibbs sampler;
2.several functions to solve the  local likelihood estimation for  exponential generalized linear model;
3.a funtions to apply the K-means algorithm.
}
\author{
Ruoxuan Li

Maintainer: Ruoxuan Li <liruoxuan@mail.ustc.edn.cn>
}
\references{
  some functions.
}
\keyword{ package }
