\name{StatComp21072-package}
\alias{StatComp21072-package}
\docType{package}
\title{
  Provide several functions for the final homework of the 'Statistical Computing' course in 2021 fall
}
\description{
  Two R functions can be used to do the Semiparametric linear regreesion with Robinson and Li method respectively.
}
\author{
Weiqi Yang.
Maintainer: Weiqi Yang <976237481@qq.com>
}
\keyword{ package }
