\name{StatComp21084-package}
\alias{StatComp21084-package}
\alias{StatComp21084}
\docType{package}
\title{
  Final project for Statistical Computation by 21084.
}
\description{
  \code{giccdmm} use coordinate descent method and multipliers for variable selection of compositional data; \code{subgroup_l0} use an approximate method of L0 penalty to perform subgroup analysis.
}
\author{
Han Wang

Maintainer: Han Wang <wh68serena@mail.ustc.edu.cn>
}
\references{
  Variable selection in regression with compositional covariates.
}
\keyword{ package }

