% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{subgroup}
\alias{subgroup}
\title{find subgroup}
\usage{
subgroup(y, x, p, m, methods)
}
\arguments{
\item{y}{the respond variable}

\item{x}{the independent variable}

\item{p}{the number of the question}

\item{m}{the number of observations}

\item{methods}{Ways to deal with missing values}
}
\value{
the unique intercept
}
\description{
find the intercept of each individual,then cluster them according to different intercepts.
}
