% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giccdmm.R
\name{subgroup_l0}
\alias{subgroup_l0}
\title{Subgroup analysis with L0 norm}
\usage{
subgroup_l0(y, X, lam)
}
\arguments{
\item{y}{the response variables}

\item{X}{the matrix of independent variables}

\item{lam}{a vector of candidate values of tuning parameters}
}
\value{
a list contains coefficients, intercept terms, BIC, and tuning parameters
}
\description{
Perform subgroup analysis with an approximate method of L0 norm
}
\examples{
\dontrun{
res <- subgroup.l0(y, X, lam)
}
}
