% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesTwo}
\alias{timesTwo}
\title{A multi-prob calculator}
\usage{
timesTwo(x, target, n)
}
\arguments{
\item{x}{the probability vector of each deliveryman}

\item{target}{there are excatly target number of deliverymen take order}

\item{n}{the length of vector x}
}
\value{
the objective probability
}
\description{
A multi-prob calculator using Rcpp
}
\examples{
\dontrun{
prob <- c(0.5, 0.5, 0.5, 0.5, 0.5)
target <- 1
n <- 5
targetoutofn(prob, target, n)
}
}
