% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Tstat}
\alias{Tstat}
\title{Use nonparametric method to test whether two random vectors are independent.}
\usage{
Tstat(x, y)
}
\arguments{
\item{x}{a random vector}

\item{y}{another random vector}
}
\value{
a scalar, T value, which can be used to test hypothesis later.
}
\description{
Use nonparametric method to test whether two random vectors are independent
by calculating the T value in the test.
}
\examples{
\dontrun{
data(faithful)
x.true <- as.numeric(faithful[,1]) # x is for eruptions
y.true <- as.numeric(faithful[,2]) # y is for waiting
alpha <- 0.05 # alpha for the test
T.true <- Tstat(x.true, y.true)
abs(T.true) > qnorm(1 - alpha / 2)
}
}
