% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tarnopolski.R
\name{Turning_point}
\alias{Turning_point}
\title{A function to compute Turning points}
\usage{
Turning_point(x)
}
\arguments{
\item{x}{A time series}
}
\value{
Turning point value
}
\description{
Calculates Turning point values.
}
\details{
Turning point of a time series.
Following Tarnopolski et al. Physica A 461 (2016) 662-673.
}
\examples{
x = arima.sim(model=list(ar = 0.3), n = 10^4)
Turning_point(x)
}
\references{
Tarnopolski et al. (2016), Physica A 461, 662-673.
}
\author{
Sebastian Sippel
}
