% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vaccC}
\alias{vaccC}
\title{Use three inputs to predict response using Rcpp.}
\usage{
vaccC(age, female, ily)
}
\arguments{
\item{age}{the first predictor (numeric)}

\item{female}{the second predictor (logical)}

\item{ily}{the third predictor (logical)}
}
\value{
a random sample of size \code{n}
}
\description{
The prediction model is described in http://www.babelgraph.org/wp/?p=358.
}
\examples{
\dontrun{
data(data)
attach(data)
res <- vaccC(age,female,ily)
}
}
