% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{count5test}
\alias{count5test}
\title{The two sample “Count Five” test for equality of variance.}
\usage{
count5test(x, y)
}
\arguments{
\item{x}{a numeric vector which represents a sample}

\item{y}{a numeric vector which represents another sample}
}
\value{
1 represents that the hypothesis is rejected, while 0 represents that the hypothesis is not rejected.
}
\description{
Suppose the means of the two samples are equal and the sample sizes are equal.
 An observation in one sample is considered extreme if it is not within the range of the other sample.
 If either sample has five or more extreme points, the hypothesis of equal variance is rejected.
}
\examples{
\dontrun{
n1 <- n2 <- 20
mu1 <- mu2 <- 0
sigma1 <- 1; sigma2 <- 2
m <- 1e3
alphahat <- mean(replicate(m,expr={
x=rnorm(n1,mu1,sigma1)
y=rnorm(n2,mu2,sigma2)
count5test(x,y)
}))
print(round(alphahat,4))
}
}
