% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{eqdist.nn}
\alias{eqdist.nn}
\title{Nearest neighbor test for equal distribution.}
\usage{
eqdist.nn(z, sizes, k, R)
}
\arguments{
\item{z}{a numeric vector which represents data sample.}

\item{sizes}{a numeric vector of length 2}

\item{k}{a scalar}

\item{R}{the number of bootstraps}
}
\value{
a list contains the bootstrap statistic and its p value
}
\description{
Nearest neighbor test for equal distribution.
}
\examples{
\dontrun{
m <- 1e2; k <- 3; p <- 2; mu <- .6
n1 <- n2 <- 20; R <- 999; N <- c(n1,n2)
p.value <- numeric(m)
for(i in 1:m){
x <- matrix(rnorm(n1*p),ncol=p)
y <- matrix(rnorm(n2*p,mean=mu),ncol=p)
z <- rbind(x,y)
p.values[i] <- eqdist.nn(z,sizes=N,k=k,R=R)$p.value
}
alpha <- 0.1
pow <- colMeans(p.values<alpha)
pow
}
}
