% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Gelman.Rubin}
\alias{Gelman.Rubin}
\title{Use the Gelman-Rubin method to monitor convergence of the chain generated by \code{rwMetropolisR}.}
\usage{
Gelman.Rubin(psi)
}
\arguments{
\item{psi}{\code{psi[i,j]} is the statistic \code{psi(X[i,1:j])}}
}
\value{
a numeric value of G-R statistic
}
\description{
Use the Gelman-Rubin method to monitor convergence of the chain
 generated by \code{rwMetropolisR}, and run the chain until it converges
 approximately to the target distribution according to \code{R_hat <1.2} .
}
\examples{
\dontrun{
 library(StatComp20006)
 k <- 4    # four chains
 x0 <- c(-10,-5,5,10)    # overdispersed initial values
 N <- 10000    # length of chains
 b <- 200    # burn-in length
 X <- matrix(nrow=k,ncol=N)
 for (i in 1:k)
   X[i,] <- rwMetropolisR(N,1,x0[i])$x
 psi <- t(apply(X, 1, cumsum))
 for (i in 1:nrow(psi))
   psi[i,] <- psi[i,] / (1:ncol(psi))
 rhat <- rep(0, N)
 for (j in (500+1):N)
   rhat[j] <- Gelman.Rubin(psi[,1:j])
 x2 <- min(which(rhat>0 & rhat<1.2))
 plot(rhat[(500+1):N], type="l", xlab="sigma=1", ylab="R_hat")
 abline(h=1.2, lty=2)
}
}
