% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler}
\usage{
gibbsC(N, thin, a, b, n)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}

\item{a}{the parameter of bivariate density}

\item{b}{the parameter of bivariate density}

\item{n}{the number of experiments}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp in Chapter 9 exercise 9.8
}
\examples{
\dontrun{
rnC <- gibbsC(100,10,2,1,10)
par(mfrow=c(2,1));
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
}
}
