% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{mylapply}
\alias{mylapply}
\title{A combination of \code{Map()} and \code{vapply()} to create an \code{lapply()} variant.}
\usage{
mylapply(X, FUN, FUN.VALUE, simplify = FALSE)
}
\arguments{
\item{X}{a vector (atomic or list) or an expression object.}

\item{FUN}{the function to be applied to each element of \code{X}}

\item{FUN.VALUE}{a (generalized) vector; a template for the return value from FUN.}

\item{simplify}{ogical or character string; should the result be simplified to a vector,
matrix or higher dimensional array if possible.}
}
\value{
a vector or a matrix of the value generated by applying FUN to the given inputs \code{X}
}
\description{
Implement a combination of \code{Map()} and \code{vapply()}
to create an \code{lapply()} variant that iterates in parallel over all of its inputs
and stores its outputs in a vector (or a matrix).
}
\examples{
\dontrun{
 data(mtcars); data(faithful)
 datalist <- list(mtcars,faithful)
 mylapply(datalist, mean, numeric(1))
}
}
