% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20065.r
\name{pwr_beta}
\alias{pwr_beta}
\title{Estimate the power of the skewness test of normality against symmetric Beta distributions.}
\usage{
pwr_beta(a)
}
\arguments{
\item{a}{parameter vector for Beta(a,a) (numeric)}
}
\value{
a list gives the power of skewness test of normality against Beta(a,a) distribution, and the standard error of the test.
}
\description{
a function from hw-10-27
}
\examples{
\dontrun{
a = c(seq(0,1,0.1),seq(1,20,1),seq(20,100,10))
pwr = pwr_beta(a)$pwr
se = pwr_beta(a)$se
plot(a, pwr, type = "b", xlab = "a", ylab = "pwr", pch=16)
abline(h = 0.1, lty = 2)
lines(a, pwr+se, lty = 4)
lines(a, pwr-se, lty = 4)
}
}
