% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsH.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{A random walk Metropolis Sampler for Generating the standard Laplace distribution}
\usage{
rw.Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation chosen for a normal distribution}

\item{x0}{the initial value of random walk}

\item{N}{size of generated sample}
}
\value{
A sample that obeys standard Laplace distribution and acceptance rates.
}
\description{
Generate  a sample that obeys standard Laplace distribution by 
Random walk Metropolis, which uses proposal distribution normal (X_t,sigma^2). 
The data generation process records the acceptance rate.
}
\examples{
\dontrun{
rw1 <- rw.Metropolis(1, 10, 5000)
plot(rw1,type="l",xlab="sigma=1",ylab="X", ylim=range(rw1))
refline <- c(log(0.05),-log(0.05))
abline(h=refline)
}
}
