\name{SA23204185-package}
\alias{SA23204185-package}
\docType{package}
\title{
  IPW and Distribution
}
\description{
  Reproduction of the IPW and distributed algorithms,a function encapsulating a simulation process as well as a calculator of a delivery problem
}
\details{
  Develop an R package to reproduce the algorithms in the paper 'Distributed learning in recommender systems,' including the IPW algorithm for recovering the right singular values of the missing matrix and the Distribution algorithm for distributed processing. We have also encapsulated the simulation process, making it easy to invoke and use.
This package includes a C++ function called targetoutofn that calculates the probability of having a specific number of deliverymen, denoted as 'target', out of a total of 'n' (where 'target' is less than or equal to 'n'). Each deliveryman has a different probability of accepting an order.
}
\author{
Zhou Runlin

Maintainer: Zhou Runlin <zrl1644@mail.ustc.edu.cn>
}
\keyword{ package }
