% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20065.r
\name{sk}
\alias{sk}
\title{a naive implement of computing skewness using R.}
\usage{
sk(x)
}
\arguments{
\item{x}{a random sample vector (numeric)}
}
\value{
a skewness of sample x
}
\description{
a function from hw-10-27 return a sample's skewness
}
\examples{
\dontrun{
x=rnorm(1000)
x.skew <- sk(x)
}
}
