% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{sk}
\alias{sk}
\title{Calculate the sample skewness}
\usage{
sk(x)
}
\arguments{
\item{x}{a numeric vector of sample}
}
\value{
a scalar of sample skewness
}
\description{
Calculate the sample skewness
}
