\name{StatComp-package}
\alias{StatComp-package}
\docType{package}
\title{
   Using dichotomy to find confidence intervals and the solution of gamma.
}
\description{
   Using dichotomy to find confidence intervals and the solution of gamma.
}
\author{
Feng Jiejing

Maintainer: Feng Jiejing <fengjj@mail.ustc.edu.cn>
}
\keyword{ package }
