\name{StatComp-package}
\alias{StatComp-package}
\docType{package}
\title{
  This package provides functions in 'Statistical Computing' course.
}
\description{
  This package include R and Rcpp functions of homework in 'Statistical Computing' course.
}
\author{
Yulin Li

Maintainer: Yulin Li <1647579023@qq.com>
}
\keyword{ package }
