\name{StatComp-package}
\alias{StatComp-package}
\docType{package}
\title{
  This package provides several functions for benchmarking R and Rcpp.
}
\description{
  R and Rcpp versions of gibbs sampler and vaac function are provided in this package (\code{gibbsR} and \code{vaccR} versus \code{gibbsC} and \code{vaccC})
}
\author{
Hong Zhang

Maintainer: Hong Zhang <zhangh@ustc.edu.cn>
}
\keyword{ package }
